+++ mod/racing_plus/Globals.lua
@@ -5,7 +5,7 @@ local g  = {}
 -- Global variables
 --
 
-g.version = "v0.47.4"
+g.version = "v0.47.5"
 g.debug = false
 g.corrupted = false -- Checked in the MC_POST_GAME_STARTED callback
 g.saveFile = { -- Checked in the MC_POST_GAME_STARTED callback
@@ -592,7 +592,7 @@ function g:ExecuteCommand(command)
   Isaac.DebugString("Finished executing command: " .. command)
 end
 
-function g:ConvertTimeToString(time)
+function g:ConvertTimeToString(time) -- time is given in seconds
   -- Calcuate the hours digit
   local hours = math.floor(time / 3600)